requirements.txt
datasets==3.1.0
einops==0.8.1
fastapi==0.120.2
geom_median==0.1.0
h5py==3.12.1
jaxtyping==0.3.3
matplotlib==3.10.0
nltk==3.9.1
numpy==1.24.0
openai==2.6.1
pandas==2.3.3
Pillow==11.1.0
plotly==6.0.0
protobuf==6.33.0
pydantic==2.12.3
python-dotenv==1.2.1
Requests==2.32.5
scikit_learn==1.7.2
scipy==1.13.1
spacy==3.8.7
streamlit==1.49.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.67.1
transformer_lens==2.9.1
transformers==4.47.1
wandb==0.22.3
uvicorn==0.34.3